move boolean_logic.py under circuits, with helper methods relocated to truth_table_oracle
