+++ qiskit/aqua/utils/__init__.py
@@ -32,40 +32,35 @@ from .qpsolver import optimize_svm
 from .circuit_factory import CircuitFactory
 from .run_circuits import compile_and_run_circuits, find_regs_by_name
 from .circuit_cache import CircuitCache
-from .boolean_logic import CNF, DNF, ESOP, get_exact_covers, get_prime_implicants
 from .backend_utils import has_ibmq, has_aer
 
 
-__all__ = ['tensorproduct',
-           'PauliGraph',
-           'convert_dict_to_json',
-           'convert_json_to_dict',
-           'random_unitary',
-           'random_h2_body',
-           'random_h1_body',
-           'random_hermitian',
-           'random_non_hermitian',
-           'decimal_to_binary',
-           'summarize_circuits',
-           'get_subsystem_density_matrix',
-           'get_subsystems_counts',
-           'get_entangler_map',
-           'validate_entangler_map',
-           'get_feature_dimension',
-           'get_num_classes',
-           'split_dataset_to_data_and_labels',
-           'map_label_to_class_name',
-           'reduce_dim_to_via_pca',
-           'optimize_svm',
-           'CircuitFactory',
-           'compile_and_run_circuits',
-           'find_regs_by_name',
-           'CircuitCache',
-           'CNF',
-           'DNF',
-           'ESOP',
-           'get_exact_covers',
-           'get_prime_implicants',
-           'has_ibmq',
-           'has_aer'
-           ]
+__all__ = [
+    'tensorproduct',
+    'PauliGraph',
+    'convert_dict_to_json',
+    'convert_json_to_dict',
+    'random_unitary',
+    'random_h2_body',
+    'random_h1_body',
+    'random_hermitian',
+    'random_non_hermitian',
+    'decimal_to_binary',
+    'summarize_circuits',
+    'get_subsystem_density_matrix',
+    'get_subsystems_counts',
+    'get_entangler_map',
+    'validate_entangler_map',
+    'get_feature_dimension',
+    'get_num_classes',
+    'split_dataset_to_data_and_labels',
+    'map_label_to_class_name',
+    'reduce_dim_to_via_pca',
+    'optimize_svm',
+    'CircuitFactory',
+    'compile_and_run_circuits',
+    'find_regs_by_name',
+    'CircuitCache',
+    'has_ibmq',
+    'has_aer',
+]
